requirements.txt
discord.py==1.7.3
httpx~=0.23.0
async-cache==1.1.1
Cython~=0.29.32
numpy==1.23.2
setuptools==64.0.3
tortoise-orm==0.19.2
asyncpg==0.26.0
